* Renamed Print* to S32, U32, and F64 * Added CRC32 check and unit test * Added command-line argument input to run_pc_terminal

+++ coretech/common/tools/communication/serial.cpp
@@ -137,7 +137,7 @@ Result Serial::Read(void * buffer, s32 bufferLength, DWORD &bytesRead)
   }
 
   bytesRead = comStat.cbInQue;
-  if(bufferLength < bytesRead)
+  if(bufferLength < static_cast<s32>(bytesRead))
     bytesRead = bufferLength;
 
   const bool fileRead = ReadFile(comPortHandle, buffer, bytesRead, &bytesRead, &readEventHandle);
